refactor: squash Alembic migrations into v1_baseline stamp
Replace 11 incremental migrations (0001-0011) with a single no-op
baseline revision. Actual schema creation handled by the init container
(Base.metadata.create_all) before Alembic runs.

Existing installations: run 'alembic stamp v1_baseline' to sync.
Fresh installations: init container creates schema then stamps.
